Substantially improved file reading functionality. Separated into a file.c that handles general IO and could ideally be used by another program, and matrix_reader.c which handles specific file reading and interpreting for matrix.c. an input2.txt contains one general example, while input.txt serves for the testing reads. Completely adapted the program to use doubles instead of floats.

+++ benchmark.c
@@ -4,9 +4,9 @@
 
 //Function that takes in another function from a library
 //and uses it to perform a shrewd benchmark
-float benchmark_general(void (*test)(void)) {
+double benchmark_general(void (*test)(void)) {
 	//Converting factor to measure in milliseconds
-	const float SECS_TO_MILLI = 1000.0;
+	const double SECS_TO_MILLI = 1000.0;
 	//Will measure time difference in milliseconds
 	double start = clock() * SECS_TO_MILLI / CLOCKS_PER_SEC;
 
@@ -18,7 +18,7 @@ float benchmark_general(void (*test)(void)) {
 	printf("\n\n");
 
 	printf("Start time: %f ms\nEnd time: %f ms\n", start, end);
-	float timeElapsed = end - start;
+	double timeElapsed = end - start;
 	printf("Time elapsed: %f ms\n", timeElapsed);
 
 	//Some output padding
@@ -29,9 +29,9 @@ float benchmark_general(void (*test)(void)) {
 
 void benchmark_general_multi(void (*test)(void), int runs) {
 	//Array holding the time entries for each time elapsed
-	float *times = malloc(runs * sizeof(float));
+	double *times = malloc(runs * sizeof(double));
 	//Total time for averaging
-	float totalTime = 0;
+	double totalTime = 0;
 
 	for(int i = 0; i < runs; i++) {
 		times[i] = benchmark_general(test);
